+++ omniqueue/src/lib.rs
@@ -87,6 +87,7 @@
 
 use std::fmt::Debug;
 
+use bytesize::ByteSize;
 use thiserror::Error;
 
 #[macro_use]
@@ -120,6 +121,15 @@ pub enum QueueError {
 
     #[error("(de)serialization error")]
     Serde(#[from] serde_json::Error),
+
+    #[error("payload too large: {} > {}", ByteSize(*actual as u64), ByteSize(*limit as u64))]
+    PayloadTooLarge {
+        /// The size of the serialized message, in bytes.
+        actual: usize,
+
+        /// The message size limit of the queue, in bytes.
+        limit: usize,
+    },
 
     #[error("{0}")]
     Generic(Box<dyn std::error::Error + Send + Sync>),