+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Remove `QueueError::Unsupported`
   - This variant was never constructed inside `omniqueue`
 
+## Additions
+
+- Add `QueueError::PayloadTooLarge`
+
 # 0.2.0
 
 This release is a big one, and we are considering omniqueue out of early development now.